Move EncryptionContext into its own file
