+++ src/main/java/com/zendesk/maxwell/row/DataJsonGenerator.java
@@ -5,8 +5,6 @@ import com.fasterxml.jackson.core.JsonGenerator;
 
 import java.io.ByteArrayOutputStream;
 import java.io.IOException;
-import java.security.NoSuchAlgorithmException;
-import java.security.SecureRandom;
 import org.apache.commons.codec.binary.Base64;
 
 interface DataJsonGenerator {
@@ -36,23 +34,6 @@ class PlaintextJsonGenerator implements DataJsonGenerator {
 
 	@Override
 	public void end(EncryptionContext ctx) throws IOException {
-	}
-}
-
-class EncryptionContext {
-	String secretKey;
-	byte[] iv;
-
-	EncryptionContext(String secretKey, byte[] iv) {
-		this.secretKey = secretKey;
-		this.iv = iv;
-	}
-
-	public static EncryptionContext create(String secretKey) throws NoSuchAlgorithmException {
-		SecureRandom randomSecureRandom = SecureRandom.getInstance("SHA1PRNG");
-		byte[] iv = new byte[16];
-		randomSecureRandom.nextBytes(iv);
-		return new EncryptionContext(secretKey, iv);
 	}
 }
